ADM-359 fix tests using store
